+++ framework/src/main/java/org/mvss/karta/framework/runtime/event/ScenarioStepCompleteEvent.java
@@ -27,12 +27,12 @@ public class ScenarioStepCompleteEvent extends Event
    private static final long serialVersionUID = 1L;
 
    private TestFeature       feature;
-   private long              iterationNumber;
+   private int               iterationNumber;
    private TestScenario      scenario;
    private TestStep          scenarioStep;
    private StepResult        result;
 
-   public ScenarioStepCompleteEvent( String runName, TestFeature feature, long iterationNumber, TestScenario scenario, TestStep scenarioStep, StepResult result )
+   public ScenarioStepCompleteEvent( String runName, TestFeature feature, int iterationNumber, TestScenario scenario, TestStep scenarioStep, StepResult result )
    {
       super( StandardEventsTypes.SCENARIO_STEP_COMPLETE_EVENT, runName );
       this.feature = feature;
@@ -43,7 +43,7 @@ public class ScenarioStepCompleteEvent extends Event
    }
 
    @Builder
-   public ScenarioStepCompleteEvent( String runName, UUID id, Date timeOfOccurrence, TestFeature feature, long iterationNumber, TestScenario scenario, TestStep scenarioStep, StepResult result )
+   public ScenarioStepCompleteEvent( String runName, UUID id, Date timeOfOccurrence, TestFeature feature, int iterationNumber, TestScenario scenario, TestStep scenarioStep, StepResult result )
    {
       super( StandardEventsTypes.SCENARIO_STEP_COMPLETE_EVENT, runName, id, timeOfOccurrence );
       this.feature = feature;
