+++ framework/src/main/java/org/mvss/karta/framework/runtime/event/ScenarioTearDownStepStartEvent.java
@@ -26,11 +26,11 @@ public class ScenarioTearDownStepStartEvent extends Event
    private static final long serialVersionUID = 1L;
 
    private TestFeature       feature;
-   private long              iterationNumber;
+   private int               iterationNumber;
    private TestScenario      scenario;
    private TestStep          scenarioTearDownStep;
 
-   public ScenarioTearDownStepStartEvent( String runName, TestFeature feature, long iterationNumber, TestScenario scenario, TestStep scenarioTearDownStep )
+   public ScenarioTearDownStepStartEvent( String runName, TestFeature feature, int iterationNumber, TestScenario scenario, TestStep scenarioTearDownStep )
    {
       super( StandardEventsTypes.SCENARIO_TEARDOWN_STEP_START_EVENT, runName );
       this.feature = feature;
@@ -40,7 +40,7 @@ public class ScenarioTearDownStepStartEvent extends Event
    }
 
    @Builder
-   public ScenarioTearDownStepStartEvent( String runName, UUID id, Date timeOfOccurrence, TestFeature feature, long iterationNumber, TestScenario scenario, TestStep scenarioTearDownStep )
+   public ScenarioTearDownStepStartEvent( String runName, UUID id, Date timeOfOccurrence, TestFeature feature, int iterationNumber, TestScenario scenario, TestStep scenarioTearDownStep )
    {
       super( StandardEventsTypes.SCENARIO_TEARDOWN_STEP_START_EVENT, runName, id, timeOfOccurrence );
       this.feature = feature;
